pyproject.toml
[tool.poetry]
name = "tz-hatikotech"
version = "0.1.0"
description = ""
authors = ["gau <gau@1jz.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.5"
djangorestframework = "^3.15.2"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pytelegrambotapi = "^4.26.0"
aiohttp = "^3.11.11"
adrf = "^0.1.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
